backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
bcrypt==4.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.37.1
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-groq==1.1.1
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
multidict==6.7.0
numpy==2.3.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyMuPDF==1.26.7
PyPDF2==3.0.1
pypdfium2==5.3.0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.13.0
uvicorn==0.37.0
xxhash==3.5.0
yarl==1.22.0
zstandard==0.25.0
